backend/requirements.txt
alembic==1.9.2
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.0.5
exceptiongroup==1.1.0
fastapi==0.89.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pydantic==1.10.4
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.10.0
python-decouple==3.7
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchgod==0.8.2
websockets==10.4
